+++ request_system/src/main/java/entity/Incident.java
@@ -1,5 +1,12 @@
 package entity;
 
+import org.hibernate.annotations.Entity;
+import org.springframework.data.annotation.Id;
+
+import javax.persistence.Column;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+
 @Entity
 public class Incident {
     @Id
@@ -8,7 +15,7 @@ public class Incident {
     private  Long id;
     private String name;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
